requirements.txt
Flask==2.3.3
pdfkit==1.0.0
python-docx==1.1.0
